+++ ChangeLog.md
@@ -10,8 +10,10 @@ and this project adheres to
 
 * Continuous integrated testing on CircleCI.
 * New Polygon class with property-based API.
-* New Polyhedron class with property-based API and robust facet sorting and merging.
 * New ConvexSpheropolygon class with property-based API.
+* New Polyhedron class with property-based API and robust facet sorting and merging.
+* New ConvexPolyhedron class with property-based API.
+* New ConvexSpheropolyhedron class with property-based API.
 
 ### Changed